refactor: reorganize feature modules
