+++ src/features/character/components/CharacterAttributes.tsx
