+++ src/features/character/legacy.ts
@@ -1,0 +1,8 @@
+import "./services/adjustments";
+import "./services/calculations";
+import "./services/characterInfo";
+import "./components/CharacterAttributes";
+import "./services/hideButtons";
+import "./services/liste";
+import "./services/skin";
+import "./services/tabs";
