+++ src/features/character/services/characterState.ts
