+++ src/features/character/services/hideButtons.ts
