+++ src/features/dice/index.ts
@@ -1,0 +1,3 @@
+export * from "./services/dice";
+export * from "./services/rechner";
+export * from "./services/spezialDice";
