+++ src/features/dice/legacy.ts
@@ -1,0 +1,3 @@
+import "./services/dice";
+import "./services/rechner";
+import "./services/spezialDice";
