+++ src/features/dice/services/dice.ts
