+++ src/features/dice/services/spezialDice.ts
