+++ src/features/shop/index.ts
@@ -1,0 +1,2 @@
+export * from "./services/shop";
+export * from "./services/wallet";
