+++ src/features/shop/legacy.ts
@@ -1,0 +1,2 @@
+import "./services/shop";
+import "./services/wallet";
